Adding support for prestodb plugin on ranger

+++ security-admin/db/sqlanywhere/patches/016-updated-schema-for-tag-based-policy.sql
@@ -153,7 +153,7 @@ CREATE TABLE dbo.x_service_resource_element_val(
 	upd_by_id bigint DEFAULT NULL NULL,
 	res_element_id bigint NOT NULL,
 	value varchar(1024) NOT NULL,
-	sort_order tinyint DEFAULT 0  NULL,
+	sort_order int DEFAULT 0  NULL,
 	CONSTRAINT x_srvc_res_el_val_PK_id PRIMARY KEY CLUSTERED(id)
 )
 GO
